Rendered some boring circles

+++ include/raytracer.h
@@ -15,7 +15,7 @@ class RayTracer {
   // Screen width and height in pixels
   size_t screen_width = 320, screen_height = 200;
   // Horizontal field of view, in degrees
-  float horizontal_fov = 60;
+  double horizontal_fov = 60;
   size_t max_depth = 5;
 
 public:
